test: use the monkeypatch builtin fixture to change to change the tmpdir
This seems a little excessive i think if there are too many
monkeypatches everywhere you wouldn't know what the appcode is supposed
to be doing, perhaps in debugging it could help but then not really use
it on production testing code?

Plus this is unnessarry wasn't there a way to change the db path through
an env var, shouldn't we just change the env var and check if the config
command confirms confirms this?